+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "rust-connectors/sources/mqtt",
     "rust-connectors/sources/http",
     "rust-connectors/sources/postgres",
+    "rust-connectors/models/fluvio-model-postgres",
     "rust-connectors/utils/mocks/http-json-mock",
     "rust-connectors/utils/fluvio-smartstream-map"
 ]
