+++ rust-connectors/models/fluvio-model-postgres/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "fluvio-model-postgres"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
