+++ rust-connectors/sources/postgres/Cargo.toml
@@ -1,25 +1,24 @@
 [package]
-name = "fluvio-connect-postgres"
-version = "0.1.0"
+name = "postgres"
+version = "0.0.0"
 edition = "2021"
 
 [[bin]]
-name = "fluvio-connect-postgres"
+name = "postgres"
 path = "src/bin/main.rs"
-required-features = ["connect"]
-
-[features]
-default = ["connect"]
-connect = ["tokio", "tokio-stream", "tokio-postgres", "postgres-protocol", "tracing", "tracing-subscriber", "fluvio"]
 
 [dependencies]
-tracing = { version = "0.1", optional = true }
-tracing-subscriber = { version = "0.3", features = ["env-filter"], optional = true }
-tokio = { version = "1", features = ["full"], optional = true }
-tokio-stream = { version = "0.1", optional = true }
-tokio-postgres = { git = "https://github.com/infinyon/rust-postgres", rev = "4e86fffc7e8d4ee10c7751a58398a172a4963bfc", optional = true }
-postgres-protocol = { git = "https://github.com/infinyon/rust-postgres", rev = "4e86fffc7e8d4ee10c7751a58398a172a4963bfc", optional = true }
-fluvio = { version = "0.9", optional = true }
+tracing = { version = "0.1" }
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+tokio = { version = "1", features = ["full"] }
+tokio-stream = { version = "0.1" }
+tokio-postgres = { git = "https://github.com/infinyon/rust-postgres", rev = "4e86fffc7e8d4ee10c7751a58398a172a4963bfc" }
+postgres-protocol = { git = "https://github.com/infinyon/rust-postgres", rev = "4e86fffc7e8d4ee10c7751a58398a172a4963bfc" }
+fluvio = { version = "0.9" }
+fluvio-dataplane-protocol = { git = "https://github.com/infinyon/fluvio", rev = "b2250033d49a279c3d718cd25c42a4cee0b40297" }
+fluvio-smartengine = { git = "https://github.com/infinyon/fluvio", rev = "b2250033d49a279c3d718cd25c42a4cee0b40297", features = ["smartmodule"] }
+fluvio-connectors-common = { path = "../../common" }
+fluvio-model-postgres = { path = "../../models/fluvio-model-postgres" }
 bytes = "1"
 eyre = "0.6"
 thiserror = "1"
